requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.7
django-cors-headers==4.3.1
django-enumchoicefield==3.0.1
django-environ==0.11.2
django-rest-enumfield==0.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
